Fixed most tests again.
