+++ client/app/components/group/groupDetail/groupDetail.spec.js
@@ -1,10 +1,10 @@
-import GroupModule from "./groupDetail";
+import GroupDetailModule from "./groupDetail";
 
 const { module } = angular.mock;
 
-describe("Group", () => {
+describe("GroupDetail", () => {
   let $httpBackend, $state, $log;
-  beforeEach(module(GroupModule));
+  beforeEach(module(GroupDetailModule));
   beforeEach(module(($stateProvider) => {
     $stateProvider
       .state("main", { url: "", abstract: true });
@@ -30,7 +30,7 @@ describe("Group", () => {
 
   describe("Module", () => {
     it("is named groupDetail", () => {
-      expect(GroupModule).to.equal("groupDetail");
+      expect(GroupDetailModule).to.equal("groupDetail");
     });
   });
 
@@ -44,13 +44,13 @@ describe("Group", () => {
     }));
 
     it("should exist", () => {
-      let $ctrl = $componentController("groupDetailComponent", {});
+      let $ctrl = $componentController("groupDetail", {});
       expect($ctrl).to.exist;
     });
 
 
     it("should be able to leave a group", () => {
-      let $ctrl = $componentController("groupDetailComponent", {});
+      let $ctrl = $componentController("groupDetail", {});
       let groupData = { id: 9834 };
       $httpBackend.expectPOST(`/api/groups/${groupData.id}/leave/`).respond(200);
       Object.assign($ctrl, { groupData });
@@ -60,7 +60,7 @@ describe("Group", () => {
     });
 
     it("clears the current group if you leave it", () => {
-      let $ctrl = $componentController("groupDetailComponent", {});
+      let $ctrl = $componentController("groupDetail", {});
       let groupData = { id: 2424 };
       $httpBackend.expectPOST(`/api/groups/${groupData.id}/leave/`).respond(200);
       CurrentGroup.set({ id: groupData.id });
@@ -73,7 +73,7 @@ describe("Group", () => {
     });
 
     it("sets an error flag if leaving fails", () => {
-      let $ctrl = $componentController("groupDetailComponent", {});
+      let $ctrl = $componentController("groupDetail", {});
       let groupData = { id: 98238 };
       $httpBackend.expectPOST(`/api/groups/${groupData.id}/leave/`).respond(400);
       Object.assign($ctrl, { groupData });
@@ -85,7 +85,8 @@ describe("Group", () => {
 
     it("saves groupData", () => {
       let groupData = { id: 667, name: "blarb" };
-      let $ctrl = $componentController("groupDetailComponent", {}, { groupData });
+      let $ctrl = $componentController("groupDetail", {});
+      $ctrl.groupData = groupData;
       let feedback = $ctrl.updateGroupData();
       $httpBackend.expectPATCH(`/api/groups/${groupData.id}/`, groupData).respond(groupData);
       $httpBackend.flush();
